feat(listener): make listener not use queue
VULN-2627
