+++ common/config.py
@@ -90,9 +90,10 @@ class BaseConfig:
         self.worker_threads = int(os.getenv("WORKER_THREADS", "0"))
         self.gunicorn_workers = int(os.getenv("GUNICORN_WORKERS", "4"))
 
+        self.max_loaded_listener_msgs = int(os.getenv("MAX_LOADED_LISTENER_MSGS", "50"))
+        self.max_loaded_grouper_msgs = int(os.getenv("MAX_LOADED_GROUPER_MSGS", "100"))
         self.max_loaded_evaluator_msgs = int(os.getenv("MAX_LOADED_EVALUATOR_MSGS", "30"))
         self.max_loaded_advisor_msgs = int(os.getenv("MAX_LOADED_ADVISOR_LISTENER_MSGS", "30"))
-        self.max_loaded_listener_msgs = int(os.getenv("MAX_LOADED_LISTENER_MSGS", "50"))
 
         self.tls_ca_path = None
         self.public_port = 8000
@@ -143,9 +144,11 @@ class BaseConfig:
         self.sasl_plain_username = None
         self.sasl_plain_password = None
 
-        # Listeners
+        # Listener
         self.listeners_valid_system_sec = int(os.getenv("LISTENERS_VALID_SYSTEM_SECONDS", "300"))
-        self.listener_messages_timeout_sec = int(os.getenv("LISTENER_MESSAGES_TIMEOUT_SECS", "10"))
+
+        # Grouper
+        self.grouper_messages_timeout_sec = int(os.getenv("GROUPER_MESSAGES_TIMEOUT_SECS", "10"))
 
         # Cluster job
         self.cluster_system_vulnerabilities = strtobool(os.getenv("CLUSTER_SYSTEM_VULNERABILITIES", "FALSE"))
@@ -204,12 +207,13 @@ class Config(BaseConfig, metaclass=Singleton):
         self.evaluator_recalc_topic = get_topic("vulnerability.evaluator.recalc")
         self.evaluator_topics = [get_topic(topic) for topic in
                                  os.getenv("EVALUATOR_TOPIC", "vulnerability.evaluator.upload,vulnerability.evaluator.recalc").split(",")]
+        self.evaluator_inventory_topic = get_topic("vulnerability.evaluator.inventory.upload")
+        self.evaluator_advisor_topic = get_topic("vulnerability.evaluator.advisor.upload")
         self.payload_tracker_topic = get_topic("platform.payload-status")
         self.advisor_results_topic = get_topic("platform.engine.results")
         self.remediation_updates_topic = get_topic("platform.remediation-updates.vulnerability")
         self.notifications_topic = get_topic("platform.notifications.ingress")
         self.evaluator_results_topic = get_topic("vulnerability.evaluator.results")
-        self.listener_ingress_topic = get_topic("vulnerability.listener.ingress")
 
         def format_uri_scheme_kafka(authtype):
             """Return protocol uri scheme"""
@@ -283,12 +287,13 @@ class Config(BaseConfig, metaclass=Singleton):
         self.evaluator_upload_topic = os.getenv("MESSAGE_TOPIC", "vulnerability.evaluator.upload")
         self.evaluator_recalc_topic = os.getenv("MESSAGE_TOPIC", "vulnerability.evaluator.recalc")
         self.evaluator_topics = os.getenv("EVALUATOR_TOPIC", "vulnerability.evaluator.upload,vulnerability.evaluator.recalc").split(",")
+        self.evaluator_inventory_topic = os.getenv("EVALUATOR_INVENTORY_TOPIC", "vulnerability.evaluator.inventory.upload")
+        self.evaluator_advisor_topic = os.getenv("EVALUATOR_ADVISOR_TOPIC", "vulnerability.evaluator.advisor.upload")
         self.payload_tracker_topic = os.getenv("PAYLOAD_TRACKER_TOPIC", "platform.payload-status")
         self.advisor_results_topic = os.getenv("ADVISOR_RESULTS_TOPIC", "platform.engine.results")
         self.remediation_updates_topic = os.getenv("REMEDIATION_UPDATES_TOPIC", "platform.remediation-updates.vulnerability")
         self.notifications_topic = os.getenv("NOTIFICATIONS_TOPIC", "platform.notifications.ingress")
         self.evaluator_results_topic = os.getenv("EVALUATOR_RESULTS_TOPIC", "vulnerability.evaluator.results")
-        self.listener_ingress_topic = os.getenv("LISTENER_INGRESS_TOPIC", "vulnerability.listener.ingress")
 
         self.rbac_url = os.environ.get("RBAC_URL")
         self.vmaas_host = os.getenv("VMAAS_HOST", "http://vmaas_webapp:8080")
