+++ common/utils.py
@@ -319,3 +319,18 @@ def send_notifications(notif_topic, new_sys_vulns, mit_sys_vulns, unmit_sys_vuln
     }
     LOGGER.debug("Sending evaluation result to notificator: %s", msg)
     notif_topic.send(msg)
+
+
+def create_task_and_log(coro, logger, loop):
+    # pylint: disable=broad-except
+    """Create asyncio task and log its result if its exception"""
+    def _log(task):
+        try:
+            task.result()
+        except asyncio.CancelledError:
+            pass
+        except Exception as exc:
+            logger.exception(exc)
+    task = loop.create_task(coro)
+    task.add_done_callback(_log)
+    return task
