+++ conf/listener_new.env
@@ -1,4 +1,3 @@
 PROMETHEUS_PORT=8089
-LISTENER_MESSAGES_TIMEOUT_SECS=10
 KAFKA_GROUP_ID=vulnerability-listener2
 MAX_LOADED_LISTENER_MSGS=50
