+++ listener/common.py
@@ -1,10 +1,9 @@
 """
 Common variables for listener components.
 """
-import asyncio
 import re
 from enum import Enum
-from datetime import datetime
+from dataclasses import dataclass
 import flags
 
 from prometheus_client import Counter
@@ -61,31 +60,17 @@ class InventoryMsgType(Enum):
     DELETE = 3
 
 
-class QueueItem:
-    """Single item in a listener queue"""
-
-    def __init__(self, inventory_msg: dict, inventory_msg_type: InventoryMsgType, advisor_msg: dict):
-        """Constructor"""
-        self.inventory_msg = inventory_msg
-        self.inventory_msg_type = inventory_msg_type
-        self.advisor_msg = advisor_msg
-        self.second_msg_event = asyncio.Event()
-        self.created = datetime.utcnow()
+@dataclass
+class InventoryMsg:
+    """Represents single inventory message"""
+    msg_type: InventoryMsgType
+    msg: dict
 
 
-def create_task_and_log(coro, logger, loop):
-    # pylint: disable=broad-except
-    """Create asyncio task and log its result if its exception"""
-    def _log(task):
-        try:
-            task.result()
-        except asyncio.CancelledError:
-            pass
-        except Exception as exc:
-            logger.exception(exc)
-    task = loop.create_task(coro)
-    task.add_done_callback(_log)
-    return task
+@dataclass
+class AdvisorMsg:
+    """Represents single advisor message"""
+    msg: dict
 
 
 async def executemany_fetchall(cur) -> []:
