+++ listener/listener.py
@@ -10,13 +10,15 @@ from psycopg_pool import AsyncConnectionPool
 
 from common.logging import init_logging, get_logger
 from common.status_app import create_status_app, create_status_runner
-from common.utils import a_ensure_minimal_schema_version, validate_kafka_msg, send_msg_to_payload_tracker
+from common.utils import (a_ensure_minimal_schema_version, validate_kafka_msg, send_msg_to_payload_tracker,
+                          create_task_and_log)
 from common.mqueue import MQReader, MQWriter
 from common.identity import get_identity, is_entitled_insights
 from common.database_handler import setup_async_db_pool
 from .common import (CFG, PROMETHEUS_PORT, SKIPPED_MESSAGES, INVALID_IDENTITY, MISSING_INSIGHTS_ENTITLEMENT,
-                     MESSAGE_PARSE_ERROR, InventoryMsgType, create_task_and_log)
-from .queue import ListenerQueue
+                     MESSAGE_PARSE_ERROR, InventoryMsgType, AdvisorMsg, InventoryMsg)
+from .advisor_processor import AdvisorMsgProcessor
+from .inventory_processor import InventoryMsgProcessor
 
 LOGGER = get_logger(__name__)
 
@@ -121,28 +123,29 @@ class Listener:
 
     def __init__(self, db_pool: AsyncConnectionPool, loop: asyncio.BaseEventLoop):
         """Constructor"""
-        self.consumer = MQReader(CFG.listener_ingress_topic)
+        self.consumer = MQReader([CFG.advisor_results_topic, CFG.events_topic])
         self.payload_tracker = MQWriter(CFG.payload_tracker_topic)
 
         self.inventory_msg_validator = InventoryMsgValidator(self.payload_tracker)
         self.advisor_msg_validator = AdvisorMsgValidator(self.payload_tracker)
 
+        self.inventory_msg_processor = InventoryMsgProcessor(db_pool, self.payload_tracker, loop)
+        self.advisor_msg_processor = AdvisorMsgProcessor(db_pool, self.payload_tracker, loop)
+
+        self.max_messages_semaphore = asyncio.BoundedSemaphore(CFG.max_loaded_listener_msgs)
         self.loop = loop
         self.db_pool = db_pool
 
-        self.queue = ListenerQueue(db_pool, self.payload_tracker, loop)
-        self._queue_task = None
-
     async def init(self):
         """Async constructor"""
-        await self.queue.init()
+        await self.inventory_msg_processor.init()
 
     def _send_payload_tracker_error(self, msg: dict, reason: str):
         # since messages come separetely, inform separetly user about error in message
         send_msg_to_payload_tracker(self.payload_tracker, msg, "received")
         send_msg_to_payload_tracker(self.payload_tracker, msg, "error", status_msg=reason)
 
-    def _consume_inventory_msg(self, msg: dict) -> InventoryMsgType:
+    async def _consume_inventory_msg(self, msg: dict) -> InventoryMsgType:
         """Consumes inventory message"""
         try:
             msg_type = self.inventory_msg_validator.validate_message(msg)
@@ -152,16 +155,19 @@ class Listener:
                 self._send_payload_tracker_error(msg, str(exc))
             except KeyError:
                 pass
-            self.queue.max_messages_semaphore.release()
             return InventoryMsgType.UNKNOWN
 
-        if msg_type is InventoryMsgType.CREATED or msg_type is InventoryMsgType.UPDATED:
-            self.queue.push_inventory_msg(msg["host"]["org_id"], msg["host"]["id"], msg, msg_type)
-        elif msg_type is InventoryMsgType.DELETE:
-            self.queue.push_inventory_msg(msg["org_id"], msg["id"], msg, msg_type)
+        await self.inventory_msg_processor.process_msg(InventoryMsg(msg_type, msg))
         return msg_type
 
-    def _consume_advisor_msg(self, msg: dict):
+    async def consume_inventory_msg(self, msg: dict):
+        """Consume inventory message, semaphore wrapper"""
+        try:
+            await self._consume_inventory_msg(msg)
+        finally:
+            self.max_messages_semaphore.release()
+
+    async def _consume_advisor_msg(self, msg: dict):
         """Consumes advisor message"""
         try:
             self.advisor_msg_validator.validate_message(msg)
@@ -171,52 +177,61 @@ class Listener:
                 self._send_payload_tracker_error(msg["input"], str(exc))
             except KeyError:
                 pass
-            self.queue.max_messages_semaphore.release()
             return
 
-        self.queue.push_advisor_msg(msg["input"]["host"]["org_id"], msg["input"]["host"]["id"], msg)
+        await self.advisor_msg_processor.process_msg(AdvisorMsg(msg))
 
-    async def _start_listening(self):
-        async for msg in self.consumer.client:
-            await self.queue.max_messages_semaphore.acquire()
-            create_task_and_log(self.consume_message(msg), LOGGER, self.loop)
+    async def consume_advisor_msg(self, msg: dict):
+        """Consume advisor message, semaphore wrapper"""
+        try:
+            await self._consume_advisor_msg(msg)
+        finally:
+            self.max_messages_semaphore.release()
 
-    async def stop(self):
-        """Stops whole listener process"""
-        LOGGER.info("Stopping the service")
-
-        await self.queue.stop()
-        await self.consumer.stop()
-        await self.payload_tracker.stop()
-        self.loop.stop()
-
-    async def run(self):
-        """Stars the listener service"""
-        await self.consumer.start()
-
-        async with self.db_pool:
-            await self.init()
-            await self._start_listening()
-            await self.stop()
-
-    async def consume_message(self, msg: ConsumerRecord):
+    def _consume_message(self, msg: ConsumerRecord):
         """Consumes single message from advisor or inventory"""
         try:
             msg_dict = json.loads(msg.value.decode("utf-8"))
         except json.decoder.JSONDecodeError as exc:
             MESSAGE_PARSE_ERROR.inc()
             LOGGER.exception("Unable to parse message: %s", exc)
-            self.queue.max_messages_semaphore.release()
             return
 
         if msg_dict.get("host") or msg_dict.get("type") == "delete":
-            self._consume_inventory_msg(msg_dict)
+            create_task_and_log(self.consume_inventory_msg(msg_dict), LOGGER, self.loop)
         elif msg_dict.get("input"):
-            self._consume_advisor_msg(msg_dict)
+            create_task_and_log(self.consume_advisor_msg(msg_dict), LOGGER, self.loop)
         else:
             LOGGER.exception("Unknown message obtained: %s", msg)
-            self.queue.max_messages_semaphore.release()
             SKIPPED_MESSAGES.inc()
+
+    async def start_listening(self):
+        """Listens for advisor/inventory messages and consumes them"""
+        async for msg in self.consumer.client:
+            await self.max_messages_semaphore.acquire()
+            self._consume_message(msg)
+
+    async def stop(self):
+        """Stops whole listener process"""
+        LOGGER.info("Stopping the service")
+
+        await self.consumer.stop()
+        await self.payload_tracker.stop()
+        await self.inventory_msg_processor.stop()
+        await self.advisor_msg_processor.stop()
+        self.loop.stop()
+
+    async def run(self):
+        """Stars the listener service"""
+        await self.consumer.start()
+        await self.inventory_msg_processor.start()
+        await self.advisor_msg_processor.start()
+        await self.payload_tracker.start()
+
+        async with self.db_pool:
+            await self.init()
+            await self.start_listening()
+            await self.stop()
 
 
 def main():
